Improve context flyout/menu sample pages
